feat: Improve cleanup of broken images in openstack
Prior to this commit if an openstack build failed it
was left behind in openstack. To stop these building up
the build won't run if there are already old version in
openstack. However the CI can't tell the difference between
images currently being built by someone else and old images.
This means that effectively only one image of a type can be run
at any one time. This is very inconvenient and effectively
elimenates the possiblility of distributed development. This feature
remedies this by cleaning up openstack properly if the builds fail.