feat: Change required arguments for create_image.py
This commit changes the template and variable file
arguments from required to optional. This is because
in nearly all the image repositories the template is called
template.json and the variable file is called variables.json. So
I changed the arguments to be optional but gave them defaults which
should work for most purposes.