fix: Add symlinks to essential tests instead of using suites
This is due to test kitchens openstack driver trying to
reuse a volume between suites. However openstack says that it is
in use and so kills the build. This is a workaround to ensure that
there is only one suite called and therefore only one machine with
one volume created.